restructure & allow override of state and setState
