+++ createActions.js
@@ -27,4 +27,3 @@ function makeAction(func, context) {
 }
 
 module.exports = createActions
-
